move write to helper.py
